Fixed bug with initialization. Fixed bug with dragging bubbles. Modularized getPageContent. Fixed bug with selecting text. Fixed error with chat outputs.
