+++ content/content.js
@@ -1,13 +1,13 @@
-import { createSidebar } from './sidebar/sidebar.js';
+import { createSidebar, getOrCreateLoadingSpinner } from './sidebar/sidebar.js';
+import { getPageContent } from './utilities/getPageContent.js';
 import { initializeModel } from './utilities/initializeModel.js';
 
 console.log("Content script loaded");
 let modelInstance = null;
-let pageContent = null;
 let modelReady = false;
 let summarizationReady = true;
 let analysisReady = true;
-let initializationReady = true;
+let initializationReady = false;
 
 // Listener for messages from popup
 chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
@@ -15,7 +15,6 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
     switch (request.action) {
         case "initializeModel":
             initModel();
-            break;
         case "showSidebar":
             createSidebar();
             break;
@@ -50,15 +49,14 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 });
 
 async function initModel(){
-    initializationReady = false;
+    initializationReady = true;
 
     // Request page content
-    pageContent = await getPageContent();
+    const pageContent = await getPageContent();
 
     // Initialize Chat Bot model
     modelInstance = await initializeModel(modelInstance, pageContent);
     modelReady = true;
-    initializationReady = true;
 }
 
 // Event listeners for updating the character count
@@ -96,87 +94,24 @@ async function summarizeContent(focusInput) {
     summarizationReady = true;
 }
 
-// Helper function to get or create a loading spinner
-function getOrCreateLoadingSpinner(parent) {
-    let loadingSpinner = document.getElementById('loadingSpinner');
-    if (!loadingSpinner) {
-        loadingSpinner = document.createElement('div');
-        loadingSpinner.id = 'loadingSpinner';
-        loadingSpinner.classList.add('spinner');
-        parent.parentElement.insertBefore(loadingSpinner, parent);
-    }
-    return loadingSpinner;
-}
-
-// Function that fetches the web page content
-async function getPageContent() {
-    let mainContent = document.querySelectorAll('article, main, section, div');
-    const contentTemp = Array.from(mainContent)
-        .filter(element => {
-            // Exclude elements with classes typically associated with non-main content
-            const className = element.className.toLowerCase();
-            if (className === '') {
-                return false;
-            }
-            const isSidebarOrNav = className.includes('sidebar') || className.includes('widget') ||
-                className.includes('related') || className.includes('nav') ||
-                className.includes('footer') || className.includes('advert') ||
-                className.includes('recirc') || className.includes('ad');
-            return !isSidebarOrNav
-        });
-
-    const contentElements = contentTemp.flatMap(element =>
-        Array.from(element.querySelectorAll('p, a, h1, h2, h3, h4, h5, h6, li, blockquote, span, figcaption, em'))
-            .filter(childElement => {
-                // Check if the current element or its parent contains 'sidebar' or any other unwanted class
-                const currentElementClass = childElement.className.toLowerCase();
-                const parentElement = childElement.parentElement;
-                const parentClass = parentElement ? parentElement.className.toLowerCase() : '';
-
-                const isCurrentElementSidebar = currentElementClass.includes('sidebar') || currentElementClass.includes('widget') ||
-                    currentElementClass.includes('related') || currentElementClass.includes('nav') ||
-                    currentElementClass.includes('footer') || currentElementClass.includes('advert') ||
-                    currentElementClass.includes('toolbar') || currentElementClass.includes('aside') ||
-                    currentElementClass.includes('ad') || currentElementClass.includes('comment') ||
-                    currentElementClass.includes('card') || currentElementClass.includes('episode') ||
-                    currentElementClass.includes('tag') || currentElementClass.includes('recommend');
-
-                const isParentSidebar = parentClass.includes('sidebar') || parentClass.includes('widget') ||
-                    parentClass.includes('related') || parentClass.includes('nav') ||
-                    parentClass.includes('footer') || parentClass.includes('advert') ||
-                    parentClass.includes('toolbar') || parentClass.includes('aside') ||
-                    parentClass.includes('ad') || parentClass.includes('comment') ||
-                    parentClass.includes('card') || parentClass.includes('episode') ||
-                    parentClass.includes('tag') || parentClass.includes('recommend');
-
-                return !isCurrentElementSidebar && !isParentSidebar; // Exclude if either the current element or its parent is a sidebar or similar
-            }));
-    const contentClean = Array.from(contentElements)
-        .map(element => {
-            let cleanedText = element.innerText.trim(); // Remove leading/trailing space
-            cleanedText = cleanedText.replace(/\s+/g, ' ').replace(/\n+/g, ' ');
-            return cleanedText.length > 0 ? cleanedText : '';
-        })
-        .filter(text => text.length > 0)
-        .filter(text => text.split(/\s+/).length >= 5);
-    const uniqueContent = Array.from(new Set(contentClean)).join('\n');
-    return uniqueContent;
-}
-
 // Function that passes page content to summarization model
 async function generateSummary(focusInput) {
+
+    // Request page content
+    const pageContent = await getPageContent();
+
     const maxChar = 3800;
     if (pageContent.length > maxChar) {
         const separateLines = pageContent.split(/\r?\n|\r|\n/g).filter(line => line.split(" ").length - 1 >= 3);
         return await summarizeLargePageContent(separateLines, maxChar, focusInput);
     } else {
 
-        return await summarizePageContent(focusInput);
+        return await summarizePageContent(pageContent, focusInput);
     }
 }
 
 // Helper function to summarize the page content in one pass
-async function summarizePageContent(focusInput) {
+async function summarizePageContent(pageContent, focusInput) {
     const summarizer = await ai.summarizer.create({ sharedContext: getSummaryContext(focusInput) });
     const summary = await summarizer.summarize(pageContent);
     summarizer.destroy();
@@ -215,7 +150,7 @@ async function summarizeLargePageContent(separateLines, maxChar, focusInput) {
 }
 
 // Function that generates the summary
-async function getSummary(summarizer, prompt, retries = 5, delay = 1000) {
+async function getSummary(summarizer, prompt, retries = 10, delay = 1000) {
     let summary = '';
     let attempt = 0;
 
@@ -255,8 +190,13 @@ function getSummaryContext(focusInput) {
 
 // Function that gets
 async function getChatBotOutput(input) {
-    const result = await modelInstance.prompt(input);
-    chrome.runtime.sendMessage({ action: "setChatBotOutput", output: result });
+    if(modelInstance){
+        const result = await modelInstance.prompt(input);
+        chrome.runtime.sendMessage({ action: "setChatBotOutput", output: result });
+    }
+    else{
+        chrome.runtime.sendMessage({ action: "setChatBotOutput", output: "Error... Model crashed..." });
+    }
 }
 
 // Function that displays fact check bubble
@@ -298,6 +238,7 @@ async function displayBubble(selectedText, type) {
 // Function to make the bubble draggable
 function makeBubbleDraggable(bubble) {
     let offsetX, offsetY;
+    let isDragging = false;
 
     bubble.addEventListener("mousedown", (e) => {
         e.preventDefault();
@@ -477,7 +418,7 @@ async function analyzeContent(pageData) {
 }
 
 // Function that analyzes web page content
-async function analyzePageText(pageData, retries = 5, delay = 1000) {
+async function analyzePageText(pageData, retries = 10, delay = 1000) {
     let result = '';
     const summary = document.getElementById('summary').textContent;
 
@@ -539,7 +480,7 @@ function getAnalysisPrompt(summary) {
 
 // Function to update the character count
 function updateCharacterCount() {
-    currentElementClass = document.getElementById("currentCharCount");
+    const currentElementClass = document.getElementById("currentCharCount");
     if (currentElementClass) {
         const selectedText = window.getSelection().toString();
         let characterCount = 0;
@@ -556,12 +497,6 @@ function updateCharacterCount() {
             }
         }
     }
-}
-
-// Function to fetch the selected content of the webpage
-function getSelectionContent() {
-    const contentElements = window.getSelection();
-    return contentElements.toString();
 }
 
 // Function to display error when analyze button is pressed and conditions are met
