+++ content/utilities/initializeModel.js
@@ -1,43 +1,40 @@
 // Function to initialize the model
 export async function initializeModel(modelInstance, pageContent) {
     try {
-        if (!modelInstance) {
+        console.log("Initializing model...");
 
-            console.log("Initializing model...");
+        const maxChar = 3800;
+        let result = null;
 
-            const maxChar = 3800;
-            let result = null;
+        // If page content exceeds maxChar, process it in chunks
+        if (pageContent.length > maxChar) {
+            let curEl = '';
+            const separateLines = pageContent.split(/\r?\n|\r|\n/g).filter(line => line.split(" ").length - 1 >= 3);
 
-            // If page content exceeds maxChar, process it in chunks
-            if (pageContent.length > maxChar) {
-                let curEl = '';
-                const separateLines = pageContent.split(/\r?\n|\r|\n/g).filter(line => line.split(" ").length - 1 >= 3);
-
-                for (const line of separateLines) {
-                    if ((curEl + line).length < maxChar) {
-                        curEl += line + '\n';
+            for (const line of separateLines) {
+                if ((curEl + line).length < maxChar) {
+                    curEl += line + '\n';
+                } else {
+                    if (!modelInstance) {
+                        modelInstance = await ai.languageModel.create({ systemPrompt: getPrompt(curEl) });
                     } else {
-                        if (!modelInstance) {
-                            modelInstance = await ai.languageModel.create({ systemPrompt: getPrompt(curEl) });
-                        } else {
-                            result = await initializeModelSection(modelInstance, curEl);
-                        }
-                        curEl = line + '\n';
+                        result = await initializeModelSection(modelInstance, curEl);
                     }
+                    curEl = line + '\n';
                 }
-
-                // Process any remaining content
-                if (curEl.trim().length > 0) {
-                    result = await initializeModelSection(modelInstance, curEl);
-                }
-            } else {
-                // Initialize the model directly with full content if within maxChar
-                modelInstance = await ai.languageModel.create({ systemPrompt: getPrompt(pageContent) });
             }
-            chrome.runtime.sendMessage({ action: "activateSendButton" });
-            chrome.runtime.sendMessage({ action: "activateSummaryButton" });
-            console.log("Model Initialized...");
-        } else { console.log("Using existing model..."); }
+
+            // Process any remaining content
+            if (curEl.trim().length > 0) {
+                result = await initializeModelSection(modelInstance, curEl);
+            }
+        } else {
+            // Initialize the model directly with full content if within maxChar
+            modelInstance = await ai.languageModel.create({ systemPrompt: getPrompt(pageContent) });
+        }
+        chrome.runtime.sendMessage({ action: "activateSendButton" });
+        chrome.runtime.sendMessage({ action: "activateSummaryButton" });
+        console.log("Model Initialized...");
     } catch (error) {
         console.error("Error initializing model:", error);
     }
@@ -45,7 +42,7 @@ export async function initializeModel(modelInstance, pageContent) {
 }
 
 // Helper function that initializes model with another section
-async function initializeModelSection(modelInstance, curEl, retries = 5, delay = 1000) {
+async function initializeModelSection(modelInstance, curEl, retries = 10, delay = 1000) {
     let result = '';
     let attempt = 0;
